processfunction/requirements.txt
boto3==1.11.13
botocore==1.14.13
Click==7.0
docutils==0.15.2
jmespath==0.9.4
Pillow==7.0.0
python-dateutil==2.8.1
python-lambda==3.2.6
PyYAML==5.3
s3transfer==0.3.3
six==1.14.0
urllib3==1.25.8
